+++ CHANGELOG.md
@@ -19,7 +19,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
-## [0.3.0.0] - 2024-03-13
+## [v0.4.0.0] - 2024-03-14
+
+### Fixed
+
+  * Handle blank selector
+
+## [v0.3.0.0] - 2024-03-14
 
 ### Added
 
@@ -47,7 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First release, a proof of concept.
 
-[unreleased]: https://github.com/someodd/gopherden/compare/v0.3.0.0...HEAD
+[unreleased]: https://github.com/someodd/gopherden/compare/v0.4.0.0...HEAD
+[0.4.0.0]: https://github.com/someodd/gopherden/compare/v0.3.0.0...v0.4.0.0
 [0.3.0.0]: https://github.com/someodd/gopherden/compare/v0.2.0.0...v0.3.0.0
 [0.2.0.0]: https://github.com/someodd/gopherden/compare/v0.1.0.0...v0.2.0.0
 [0.1.0.0]: https://github.com/someodd/gopherden/release/v0.1.0.0